Fixed issue while saving a module with multilingual rich text box.
Voyager 1.2.3 have an issue with saving/editing multilingual rich text box.

The root cause of this issue is due to the position of DOM element for rich text box inputs. This changes will fix the issue by first targeting the parents of the "_il8n" element and then search for the input field with the class name "form-control".